refactor: 💡 repository function of teamPlan
